Dual pipelines working. Still bugs.

+++ ui/qt/common/QtChartBase.h
@@ -24,7 +24,7 @@ public:
   virtual void plot(const vsdrreal& data, uint32_t length);
   virtual void plot(const vsdrcomplex data, uint32_t length);
 
-  void setSeriesXMinMax(uint32_t min, uint32_t max);
+  void setSeriesXMinMax(int64_t min, int64_t max);
 
 protected:
   virtual void applyTheme();
@@ -37,6 +37,6 @@ protected:
   QChart* m_pChart;
   QtChartTheme* m_pTheme;
   QLineSeries m_lineSeries;
-  uint32_t m_xMin;
-  uint32_t m_xMax;
+  int64_t m_xMin;
+  int64_t m_xMax;
 };
